chore: remove dead admin handlers and unused exports
The /sync-logs, /account-links, /webhook-events, /mcp-guide,
/mcp-settings, and /agent-wizard pages were absorbed into tab-based
pages with redirects in router.go but the old handlers and templates
remained. Removed the dead handlers, their standalone templates, and
helper functions found via the deadcode tool:

- admin: SyncLogsHandler, AccountLinksPageHandler, WebhookEventsHandler,
  MCPGuideHandler, MCPSettingsGetHandler, AgentWizardHandler, the
  generateState OAuth helper, and stale RegisterBasePage/RenderPartial/
  RenderTo template helpers
- mcp: unused MCPServer.Server() (BuildServer is canonical)
- prompts: AllAgentTypes
- sync: RuleResolver.UncategorizedID
- templates/components: AvatarURLVersioned

Kept the agents_page composite extras (mcp_settings/agent_wizard/
mcp_guide HTML now used as embedded blocks in agents.html), and kept
AccountForLink which connections.go consumes. Updated rules_render_test
to use httptest.NewRecorder + tr.Render in place of the removed RenderTo.

Net -1294 LOC.
