+++ internal/admin/mcp.go
@@ -10,156 +10,22 @@ import (
 	"github.com/alexedwards/scs/v2"
 )
 
-// MCPSettingsGetHandler serves GET /admin/mcp.
-func MCPSettingsGetHandler(svc *service.Service, mcpServer *breadboxmcp.MCPServer, sm *scs.SessionManager, tr *TemplateRenderer) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		cfg, err := svc.GetMCPConfig(r.Context())
-		if err != nil {
-			http.Error(w, "Failed to load MCP config", http.StatusInternalServerError)
-			return
-		}
-
-		// Build tool list for display.
-		type toolInfo struct {
-			Name           string
-			Description    string
-			Classification string
-			Enabled        bool
-			Group          string
-		}
-		disabledSet := make(map[string]bool)
-		for _, t := range cfg.DisabledTools {
-			disabledSet[t] = true
-		}
-
-		// Tool group assignments (order matters for display).
-		toolGroups := map[string]string{
-			"list_accounts":                "Accounts & Data",
-			"list_users":                   "Accounts & Data",
-			"get_sync_status":              "Accounts & Data",
-			"trigger_sync":                 "Accounts & Data",
-			"query_transactions":           "Transactions",
-			"count_transactions":           "Transactions",
-			"transaction_summary":          "Transactions",
-			"merchant_summary":             "Transactions",
-			"list_categories":              "Categories",
-			"export_categories":            "Categories",
-			"import_categories":            "Categories",
-			"categorize_transaction":       "Categorization",
-			"reset_transaction_category":   "Categorization",
-			"batch_categorize_transactions": "Categorization",
-			"bulk_recategorize":            "Categorization",
-			"list_tags":                    "Tags",
-			"create_tag":                   "Tags",
-			"update_tag":                   "Tags",
-			"delete_tag":                   "Tags",
-			"add_transaction_tag":          "Tags",
-			"remove_transaction_tag":       "Tags",
-			"update_transactions":          "Tags",
-			"list_annotations":             "Tags",
-			"list_transaction_rules":       "Rules",
-			"create_transaction_rule":      "Rules",
-			"update_transaction_rule":      "Rules",
-			"delete_transaction_rule":      "Rules",
-			"batch_create_rules":           "Rules",
-			"apply_rules":                 "Rules",
-			"preview_rule":                "Rules",
-			"list_account_links":           "Account Links",
-			"create_account_link":          "Account Links",
-			"delete_account_link":          "Account Links",
-			"reconcile_account_link":       "Account Links",
-			"list_transaction_matches":     "Account Links",
-			"confirm_match":               "Account Links",
-			"reject_match":                "Account Links",
-			"add_transaction_comment":      "Comments & Reports",
-			"list_transaction_comments":    "Comments & Reports",
-			"submit_report":               "Comments & Reports",
-		}
-		// Ordered group names for display.
-		groupOrder := []string{
-			"Accounts & Data", "Transactions", "Categories", "Categorization",
-			"Tags", "Rules", "Account Links", "Comments & Reports",
-		}
-
-		// Build tools grouped in order.
-		toolsByGroup := make(map[string][]toolInfo)
-		for _, td := range mcpServer.AllToolDefs() {
-			enabled := !disabledSet[td.Tool.Name]
-			group := toolGroups[td.Tool.Name]
-			if group == "" {
-				group = "Other"
-			}
-			toolsByGroup[group] = append(toolsByGroup[group], toolInfo{
-				Name:           td.Tool.Name,
-				Description:    td.Tool.Description,
-				Classification: string(td.Classification),
-				Enabled:        enabled,
-				Group:          group,
-			})
-		}
-		// Flatten in group order.
-		var tools []toolInfo
-		for _, g := range groupOrder {
-			tools = append(tools, toolsByGroup[g]...)
-		}
-		if others := toolsByGroup["Other"]; len(others) > 0 {
-			tools = append(tools, others...)
-		}
-
-		// If no saved instructions, show the defaults.
-		instructions := cfg.Instructions
-		if instructions == "" {
-			instructions = breadboxmcp.DefaultInstructions
-		}
-
-		enabledCount := 0
-		for _, t := range tools {
-			if t.Enabled {
-				enabledCount++
-			}
-		}
-
-		// Build grouped tool data for template.
-		type toolGroup struct {
-			Name  string
-			Tools []toolInfo
-		}
-		var toolGroupList []toolGroup
-		for _, g := range groupOrder {
-			if ts := toolsByGroup[g]; len(ts) > 0 {
-				toolGroupList = append(toolGroupList, toolGroup{Name: g, Tools: ts})
-			}
-		}
-		if others := toolsByGroup["Other"]; len(others) > 0 {
-			toolGroupList = append(toolGroupList, toolGroup{Name: "Other", Tools: others})
-		}
-
-		// Resolve review guidelines and report format (custom or default).
-		reviewGuidelines := cfg.ReviewGuidelines
-		if reviewGuidelines == "" {
-			reviewGuidelines = breadboxmcp.DefaultReviewGuidelines
-		}
-		reportFormat := cfg.ReportFormat
-		if reportFormat == "" {
-			reportFormat = breadboxmcp.DefaultReportFormat
-		}
-
-		data := BaseTemplateData(r, sm, "mcp", "MCP Settings")
-		data["MCPConfig"] = cfg
-		data["Tools"] = tools
-		data["ToolGroups"] = toolGroupList
-		data["ToolsEnabledCount"] = enabledCount
-		data["ToolsDisabledCount"] = len(tools) - enabledCount
-		data["ToolsTotalCount"] = len(tools)
-		data["Instructions"] = instructions
-		data["DefaultInstructions"] = breadboxmcp.DefaultInstructions
-		data["ReviewGuidelines"] = reviewGuidelines
-		data["DefaultReviewGuidelines"] = breadboxmcp.DefaultReviewGuidelines
-		data["ReportFormat"] = reportFormat
-		data["DefaultReportFormat"] = breadboxmcp.DefaultReportFormat
-
-		tr.Render(w, r, "mcp_settings.html", data)
+// mcpServerURL derives the public MCP endpoint from the incoming request,
+// honoring X-Forwarded-Proto / X-Forwarded-Host when set by a reverse proxy.
+// Shared across admin pages that surface the URL to users.
+func mcpServerURL(r *http.Request) string {
+	scheme := "http"
+	if r.TLS != nil {
+		scheme = "https"
 	}
+	if proto := r.Header.Get("X-Forwarded-Proto"); proto != "" {
+		scheme = proto
+	}
+	host := r.Host
+	if fwdHost := r.Header.Get("X-Forwarded-Host"); fwdHost != "" {
+		host = fwdHost
+	}
+	return scheme + "://" + host + "/mcp"
 }
 
 // MCPSaveModeHandler handles POST /admin/mcp/mode.
