+++ internal/admin/synclogs.go
@@ -11,120 +11,6 @@ import (
 	"github.com/go-chi/chi/v5"
 )
 
-// SyncLogsHandler serves GET /admin/sync-logs.
-func SyncLogsHandler(a *app.App, sm *scs.SessionManager, tr *TemplateRenderer, svc *service.Service) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		ctx := r.Context()
-
-		// Parse query params.
-		q := r.URL.Query()
-		params := service.SyncLogListParams{
-			Page:         parsePage(r),
-			PageSize:     25,
-			ConnectionID: optStrQuery(q, "connection_id"),
-			Status:       optStrQuery(q, "status"),
-			Trigger:      optStrQuery(q, "trigger"),
-			DateFrom:     parseDateParam(r, "date_from"),
-			DateTo:       parseInclusiveDateParam(r, "date_to"),
-		}
-
-		// Fetch paginated sync logs.
-		result, err := svc.ListSyncLogsPaginated(ctx, params)
-		if err != nil {
-			a.Logger.Error("list sync logs", "error", err)
-			http.Error(w, "Internal server error", http.StatusInternalServerError)
-			return
-		}
-
-		// Fetch connections for the filter dropdown.
-		connections, err := a.Queries.ListBankConnections(ctx)
-		if err != nil {
-			a.Logger.Error("list bank connections for sync log filters", "error", err)
-		}
-
-		filterConnID := stringOrEmpty(params.ConnectionID)
-		filterStatus := stringOrEmpty(params.Status)
-		filterTrigger := stringOrEmpty(params.Trigger)
-		filterDateFrom := r.URL.Query().Get("date_from")
-		filterDateTo := r.URL.Query().Get("date_to")
-
-		// Build base params for stats/counts (non-status filters apply, status does not).
-		baseParams := service.SyncLogListParams{
-			ConnectionID: params.ConnectionID,
-			Trigger:      params.Trigger,
-			DateFrom:     params.DateFrom,
-			DateTo:       params.DateTo,
-		}
-
-		stats, err := svc.SyncLogStats(ctx, baseParams)
-		if err != nil {
-			a.Logger.Error("sync log stats", "error", err)
-			stats = &service.SyncLogStats{} // graceful fallback
-		}
-
-		// Count per-status for the tab badges (include trigger+date but not status).
-		successStatus := "success"
-		errorStatus := "error"
-		inProgressStatus := "in_progress"
-		successParams := baseParams
-		successParams.Status = &successStatus
-		errorParams := baseParams
-		errorParams.Status = &errorStatus
-		inProgressParams := baseParams
-		inProgressParams.Status = &inProgressStatus
-
-		successCount, _ := svc.CountSyncLogsFiltered(ctx, successParams)
-		errorCount, _ := svc.CountSyncLogsFiltered(ctx, errorParams)
-		inProgressCount, _ := svc.CountSyncLogsFiltered(ctx, inProgressParams)
-
-		// Determine whether the advanced filter panel has active filters.
-		hasAdvancedFilters := filterConnID != "" || filterTrigger != "" || filterDateFrom != "" || filterDateTo != ""
-
-		pageSize := 25
-		showingStart := (result.Page-1)*pageSize + 1
-		showingEnd := result.Page * pageSize
-		if int64(showingEnd) > result.Total {
-			showingEnd = int(result.Total)
-		}
-
-		pBase := paginationBase("/sync-logs", map[string]string{
-			"connection_id": filterConnID,
-			"status":        filterStatus,
-			"trigger":       filterTrigger,
-			"date_from":     filterDateFrom,
-			"date_to":       filterDateTo,
-		}, "page")
-
-		data := map[string]any{
-			"PageTitle":          "Sync Logs",
-			"CurrentPage":        "sync-logs",
-			"CSRFToken":          GetCSRFToken(r),
-			"Flash":              GetFlash(ctx, sm),
-			"Logs":               result.Logs,
-			"Connections":        connections,
-			"FilterConnID":       filterConnID,
-			"FilterStatus":       filterStatus,
-			"FilterTrigger":      filterTrigger,
-			"FilterDateFrom":     filterDateFrom,
-			"FilterDateTo":       filterDateTo,
-			"HasAdvancedFilters": hasAdvancedFilters,
-			"Page":               result.Page,
-			"TotalPages":         result.TotalPages,
-			"Total":              result.Total,
-			"PageSize":           pageSize,
-			"ShowingStart":       showingStart,
-			"ShowingEnd":         showingEnd,
-			"PaginationBase":     pBase,
-			"Stats":              stats,
-			"SuccessCount":       successCount,
-			"ErrorCount":         errorCount,
-			"InProgressCount":    inProgressCount,
-			"WarningCount":       stats.WarningCount,
-		}
-		tr.Render(w, r, "sync_logs.html", data)
-	}
-}
-
 // SyncLogDetailHandler serves GET /admin/sync-logs/{id}.
 func SyncLogDetailHandler(a *app.App, sm *scs.SessionManager, tr *TemplateRenderer, svc *service.Service) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
