--- internal/templates/pages/account_links.html
@@ -1,134 +1,0 @@
-{{template "base" .}}
-
-{{define "content"}}
-<div class="bb-page-header">
-  <div>
-    <h1 class="bb-page-title">Account Links</h1>
-    <p class="text-base-content/60 text-sm mt-1">Link dependent (authorized user) accounts to primary accounts for transaction deduplication and attribution.</p>
-  </div>
-  <button class="btn btn-primary btn-sm rounded-xl" onclick="document.getElementById('create-modal').showModal()">
-    <i data-lucide="plus" class="w-4 h-4"></i> New Link
-  </button>
-</div>
-
-{{if .Links}}
-<div class="space-y-3 bb-stagger">
-  {{range .Links}}
-  <div class="bb-card p-5">
-    <div class="flex flex-col sm:flex-row sm:items-start justify-between gap-3">
-      <div class="flex-1 min-w-0">
-        <!-- Account link: stacked on mobile, inline on desktop -->
-        <div class="flex flex-col sm:flex-row sm:items-center gap-1 sm:gap-2">
-          <div class="flex items-center gap-2 min-w-0">
-            <div class="flex items-center justify-center w-8 h-8 rounded-lg bg-primary/8 shrink-0">
-              <i data-lucide="credit-card" class="w-4 h-4 text-primary"></i>
-            </div>
-            <div class="min-w-0">
-              <span class="font-semibold text-sm truncate block">{{.PrimaryAccountName}}</span>
-              <span class="text-xs text-base-content/40">{{.PrimaryUserName}}</span>
-            </div>
-          </div>
-          <i data-lucide="arrow-right" class="w-4 h-4 text-base-content/30 shrink-0 hidden sm:block"></i>
-          <i data-lucide="arrow-down" class="w-4 h-4 text-base-content/30 shrink-0 sm:hidden ml-3"></i>
-          <div class="flex items-center gap-2 min-w-0 sm:ml-0 ml-3">
-            <div class="flex items-center justify-center w-8 h-8 rounded-lg bg-secondary/8 shrink-0">
-              <i data-lucide="credit-card" class="w-4 h-4 text-secondary"></i>
-            </div>
-            <div class="min-w-0">
-              <span class="font-semibold text-sm truncate block">{{.DependentAccountName}}</span>
-              <span class="text-xs text-base-content/40">{{.DependentUserName}}</span>
-            </div>
-          </div>
-          {{if not .Enabled}}<span class="badge badge-ghost badge-sm">Disabled</span>{{end}}
-        </div>
-        <!-- Stats row -->
-        <div class="flex items-center gap-3 sm:gap-4 mt-3 text-xs text-base-content/50 flex-wrap">
-          <span class="flex items-center gap-1">
-            <i data-lucide="check-circle-2" class="w-3.5 h-3.5 text-success"></i>
-            {{.MatchCount}} matched
-          </span>
-          {{if gt .UnmatchedDependentCount 0}}
-          <span class="flex items-center gap-1">
-            <i data-lucide="help-circle" class="w-3.5 h-3.5 text-warning"></i>
-            {{.UnmatchedDependentCount}} unmatched
-          </span>
-          {{end}}
-          <span class="hidden sm:inline">Strategy: {{.MatchStrategy}}</span>
-          {{if gt .MatchToleranceDays 0}}<span class="hidden sm:inline">Tolerance: {{.MatchToleranceDays}}d</span>{{end}}
-        </div>
-      </div>
-      <!-- Actions -->
-      <div class="flex items-center gap-1 shrink-0 sm:self-start self-end">
-        <form method="POST" action="/-/account-links/{{.ID}}/reconcile" class="inline">
-          <input type="hidden" name="_csrf" value="{{$.CSRFToken}}">
-          <button type="submit" class="btn btn-ghost btn-xs rounded-lg" title="Run reconciliation">
-            <i data-lucide="refresh-cw" class="w-3.5 h-3.5"></i>
-          </button>
-        </form>
-        <a href="/account-links/{{.ID}}" class="btn btn-ghost btn-xs rounded-lg" title="View matches">
-          <i data-lucide="list" class="w-3.5 h-3.5"></i>
-        </a>
-        <form method="POST" action="/-/account-links/{{.ID}}/delete" class="inline"
-              x-data @submit.prevent="bbConfirm({title:'Delete account link?',message:'All transaction attribution from this link will be cleared. This cannot be undone.',confirmLabel:'Delete',variant:'danger'}).then(ok => { if(ok) $el.submit(); })">
-          <input type="hidden" name="_csrf" value="{{$.CSRFToken}}">
-          <button type="submit" class="btn btn-ghost btn-xs text-error rounded-lg" title="Delete link">
-            <i data-lucide="trash-2" class="w-3.5 h-3.5"></i>
-          </button>
-        </form>
-      </div>
-    </div>
-  </div>
-  {{end}}
-</div>
-{{else}}
-<div class="bb-card p-12 text-center">
-  <div class="flex flex-col items-center">
-    <div class="w-16 h-16 rounded-xl bg-base-200/60 flex items-center justify-center mb-4">
-      <i data-lucide="link-2" class="w-8 h-8 text-base-content/25"></i>
-    </div>
-    <h3 class="text-lg font-semibold mb-1">No Account Links</h3>
-    <p class="text-sm text-base-content/50 mb-5 max-w-sm">
-      Account links connect authorized user accounts to primary accounts for deduplication.
-      If two family members connect the same credit card, linking prevents double-counting.
-    </p>
-    <button class="btn btn-primary btn-sm rounded-xl" onclick="document.getElementById('create-modal').showModal()">
-      <i data-lucide="plus" class="w-4 h-4"></i> Create First Link
-    </button>
-  </div>
-</div>
-{{end}}
-
-<!-- Create Link Modal -->
-<dialog id="create-modal" class="modal modal-bottom sm:modal-middle">
-  <div class="modal-box rounded-xl max-w-lg">
-    <h3 class="font-bold text-lg">Create Account Link</h3>
-    <p class="text-sm text-base-content/50 mt-1">Link a dependent (authorized user) account to the primary cardholder's account.</p>
-    <form method="POST" action="/-/account-links" class="mt-4 space-y-4" x-data="{ loading: false }" @submit="loading = true">
-      <input type="hidden" name="_csrf" value="{{.CSRFToken}}">
-      <div class="form-control">
-        <label class="label label-text text-sm font-medium">Primary Account</label>
-        <select name="primary_account_id" class="select select-bordered select-sm w-full rounded-xl" required>
-          <option value="">Select the primary cardholder's account...</option>
-          {{range .Accounts}}
-          <option value="{{.ID}}">{{.DisplayName}}{{if .Mask}} ••{{.Mask}}{{end}} ({{.UserName}} - {{.InstitutionName}})</option>
-          {{end}}
-        </select>
-      </div>
-      <div class="form-control">
-        <label class="label label-text text-sm font-medium">Dependent Account</label>
-        <select name="dependent_account_id" class="select select-bordered select-sm w-full rounded-xl" required>
-          <option value="">Select the authorized user's account...</option>
-          {{range .Accounts}}
-          <option value="{{.ID}}">{{.DisplayName}}{{if .Mask}} ••{{.Mask}}{{end}} ({{.UserName}} - {{.InstitutionName}})</option>
-          {{end}}
-        </select>
-      </div>
-      <div class="modal-action">
-        <button type="button" class="btn btn-ghost btn-sm rounded-xl" onclick="document.getElementById('create-modal').close()">Cancel</button>
-        <button type="submit" class="btn btn-primary btn-sm rounded-xl" :class="{ 'loading': loading }" :disabled="loading">Create Link</button>
-      </div>
-    </form>
-  </div>
-  <form method="dialog" class="modal-backdrop"><button>close</button></form>
-</dialog>
-{{end}}
