+++ internal/templates/pages/agent_wizard.html
@@ -1,5 +1,3 @@
-{{define "content"}}{{template "agents-wizard-content" .}}{{end}}
-
 {{define "agents-wizard-content"}}
 {{$stats := .Stats}}
 <div role="alert" class="alert alert-info alert-soft text-sm rounded-xl mb-6">
