+++ internal/templates/pages/mcp_guide.html
@@ -1,5 +1,3 @@
-{{define "content"}}{{template "agents-guide-content" .}}{{end}}
-
 {{define "agents-guide-content"}}
 
 <!-- Steps indicator -->
