+++ internal/templates/pages/mcp_settings.html
@@ -1,5 +1,3 @@
-{{define "content"}}{{template "agents-settings-content" .}}{{end}}
-
 {{define "agents-settings-content"}}
 <div class="grid gap-6 bb-stagger">
